requirements.txt
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
schedule>=1.2.0
python-Levenshtein>=0.21.0
openpyxl>=3.1.0
aiohttp>=3.8.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0 